sdk-support-copilot/requirements.txt
openai
openai-agents
python-dotenv
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
mcp
uvicorn
starlette